feat(common): refactor common package
