+++ packages/ceramic-common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ceramic-common",
+  "name": "@ceramicnetwork/ceramic-common",
   "version": "1.0.0",
   "description": "Ceramic common types and utilities",
   "keywords": [
