+++ packages/ceramic-core/src/doctype/three-id/three-id-doctype-handler.ts
@@ -8,8 +8,8 @@ import { verifyJWT } from 'did-jwt'
 import { ThreeIdDoctype, ThreeIdParams } from "./three-id-doctype"
 import {
     AnchorProof, AnchorRecord, AnchorStatus, DocState, DoctypeConstructor, DoctypeHandler, InitOpts, SignatureStatus
-} from "ceramic-common/lib/doctype"
-import { Context } from "ceramic-common/lib/context"
+} from "@ceramicnetwork/ceramic-common/lib/doctype"
+import { Context } from "@ceramicnetwork/ceramic-common/lib/context"
 
 const DOCTYPE = '3id'
