+++ packages/ceramic-core/src/store/__tests__/level-state-store.test.ts
@@ -112,9 +112,9 @@ jest.mock("../../ceramic-user")
 import CeramicUser from "../../ceramic-user"
 import ThreeIdDoctypeHandler from "../../doctype/three-id/three-id-doctype-handler"
 import { ThreeIdDoctype } from "../../doctype/three-id/three-id-doctype"
-import { Doctype, DoctypeHandler } from "../../../../ceramic-common/lib/doctype"
-import AnchorService from "../../../../ceramic-common/lib/anchor-service"
-import { Context } from "../../../../ceramic-common/lib/context"
+import { Doctype, DoctypeHandler } from "../../../../@ceramicnetwork/ceramic-common/lib/doctype"
+import AnchorService from "../../../../@ceramicnetwork/ceramic-common/lib/anchor-service"
+import { Context } from "../../../../@ceramicnetwork/ceramic-common/lib/context"
 
 jest.mock("did-jwt", () => ({
   verifyJWT: (): any => 'verified'
